test_app/requirement.txt
argparse==1.2.1
Django==1.7.4
djangorestframework==3.0.4
MySQL-python==1.2.5
pip==1.5.6
setuptools==3.6
six==1.9.0
wsgiref==0.1.2
